# No Dynamic RTree for Polygon Validation