## Description

Uses a bulk-load approach rather than a dynamic approach with the RTree for Polygon validation.

This doesn't give us a performance improvement (there is no real performance change). However, it paves the way for removal of the RTree Delete and Insert methods, which contain considerable complexity.

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- N/A

## Benchmark Results

<details>

<summary>Click to expand</summary>

```

name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.60ns ± 7%    1.62ns ±14%     ~     (p=0.829 n=13+15)
LineEnvelope/1-4                                              1.06ns ± 6%    1.06ns ± 7%     ~     (p=0.846 n=14+14)
LineEnvelope/2-4                                              1.30ns ± 6%    1.33ns ±15%     ~     (p=0.140 n=13+13)
LineEnvelope/3-4                                              1.19ns ±10%    1.20ns ±15%     ~     (p=0.991 n=14+15)
MarshalWKB/polygon/n=10-4                                      190ns ±14%     199ns ±12%     ~     (p=0.095 n=15+15)
MarshalWKB/polygon/n=100-4                                     562ns ±10%     560ns ±14%     ~     (p=0.821 n=14+15)
MarshalWKB/polygon/n=1000-4                                   3.83µs ±21%    3.72µs ±20%     ~     (p=0.480 n=15+15)
MarshalWKB/polygon/n=10000-4                                  31.5µs ±26%    34.9µs ±25%     ~     (p=0.072 n=13+15)
UnmarshalWKB/polygon/n=10-4                                    314ns ± 7%     319ns ±11%     ~     (p=0.788 n=14+15)
UnmarshalWKB/polygon/n=100-4                                   716ns ±15%     697ns ± 9%     ~     (p=0.484 n=15+14)
UnmarshalWKB/polygon/n=1000-4                                 4.20µs ±27%    4.09µs ±17%     ~     (p=0.806 n=15+15)
UnmarshalWKB/polygon/n=10000-4                                37.0µs ±30%    37.3µs ±15%     ~     (p=0.652 n=15+14)
IntersectsLineStringWithLineString/n=10-4                     1.49µs ± 6%    1.48µs ±15%     ~     (p=0.464 n=12+15)
IntersectsLineStringWithLineString/n=100-4                    21.2µs ±11%    20.1µs ± 7%   -5.13%  (p=0.037 n=15+14)
IntersectsLineStringWithLineString/n=1000-4                    207µs ±10%     204µs ±10%     ~     (p=0.401 n=14+14)
IntersectsLineStringWithLineString/n=10000-4                  3.09ms ±19%    2.90ms ± 7%     ~     (p=0.085 n=15+14)
IntersectsMultiPointWithMultiPoint/n=20-4                     1.34µs ± 9%    1.31µs ± 5%   -2.51%  (p=0.032 n=14+13)
IntersectsMultiPointWithMultiPoint/n=200-4                    14.1µs ± 6%    13.8µs ± 5%     ~     (p=0.093 n=15+14)
IntersectsMultiPointWithMultiPoint/n=2000-4                    140µs ± 9%     138µs ± 3%     ~     (p=0.347 n=13+12)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.50ms ± 4%    1.51ms ± 7%     ~     (p=0.979 n=12+13)
LineStringIsSimpleZigZag/10-4                                 1.90µs ±17%    1.91µs ±13%     ~     (p=0.783 n=15+15)
LineStringIsSimpleZigZag/100-4                                29.8µs ±13%    30.0µs ±12%     ~     (p=0.603 n=14+14)
LineStringIsSimpleZigZag/1000-4                                363µs ± 9%     356µs ± 4%     ~     (p=0.285 n=14+14)
LineStringIsSimpleZigZag/10000-4                              4.81ms ± 7%    5.08ms ±28%     ~     (p=0.847 n=14+15)
PolygonSingleRingValidation/n=10-4                            2.37µs ± 6%    2.54µs ±18%   +7.03%  (p=0.014 n=15+14)
PolygonSingleRingValidation/n=100-4                           32.0µs ±12%    32.0µs ± 8%     ~     (p=0.839 n=14+14)
PolygonSingleRingValidation/n=1000-4                           389µs ± 9%     392µs ±13%     ~     (p=0.769 n=14+14)
PolygonSingleRingValidation/n=10000-4                         4.90ms ± 9%    5.07ms ±22%     ~     (p=0.949 n=14+15)
PolygonMultipleRingsValidation/n=4-4                          6.61µs ± 6%    6.87µs ±19%     ~     (p=0.614 n=12+15)
PolygonMultipleRingsValidation/n=36-4                         60.0µs ±13%    58.1µs ±15%     ~     (p=0.148 n=15+15)
PolygonMultipleRingsValidation/n=400-4                         770µs ±10%     720µs ± 6%   -6.53%  (p=0.000 n=14+12)
PolygonMultipleRingsValidation/n=4096-4                       9.07ms ± 6%    8.62ms ± 8%   -4.91%  (p=0.005 n=14+14)
PolygonZigZagRingsValidation/n=10-4                           11.0µs ± 6%    11.0µs ± 7%     ~     (p=0.946 n=14+14)
PolygonZigZagRingsValidation/n=100-4                           129µs ±11%     127µs ± 8%     ~     (p=0.511 n=14+14)
PolygonZigZagRingsValidation/n=1000-4                         1.45ms ±11%    1.45ms ± 5%     ~     (p=0.482 n=14+14)
PolygonZigZagRingsValidation/n=10000-4                        18.9ms ±13%    18.8ms ± 5%     ~     (p=0.892 n=15+13)
PolygonAnnulusValidation/n=10-4                               3.65µs ±17%    3.74µs ±10%     ~     (p=0.093 n=14+15)
PolygonAnnulusValidation/n=100-4                              32.0µs ± 6%    33.0µs ± 9%     ~     (p=0.065 n=13+15)
PolygonAnnulusValidation/n=1000-4                              549µs ±10%     543µs ± 7%     ~     (p=0.550 n=13+14)
PolygonAnnulusValidation/n=10000-4                            6.51ms ±10%    6.54ms ±11%     ~     (p=0.983 n=14+15)
MultipolygonValidation/n=1-4                                   299ns ±16%     300ns ± 8%     ~     (p=0.315 n=15+14)
MultipolygonValidation/n=4-4                                   730ns ±10%     715ns ± 8%     ~     (p=0.280 n=14+14)
MultipolygonValidation/n=16-4                                 4.31µs ±11%    4.22µs ± 9%     ~     (p=0.560 n=15+15)
MultipolygonValidation/n=64-4                                 22.3µs ± 8%    22.4µs ±11%     ~     (p=0.847 n=15+14)
MultipolygonValidation/n=256-4                                 158µs ± 4%     157µs ± 3%     ~     (p=0.160 n=14+12)
MultipolygonValidation/n=1024-4                                755µs ±10%     742µs ± 6%     ~     (p=0.581 n=15+12)
MultiPolygonTwoCircles/n=10-4                                 3.56µs ± 9%    3.54µs ±16%     ~     (p=0.367 n=15+15)
MultiPolygonTwoCircles/n=100-4                                37.6µs ±13%    37.2µs ± 7%     ~     (p=0.981 n=15+12)
MultiPolygonTwoCircles/n=1000-4                                406µs ±10%     412µs ±11%     ~     (p=0.451 n=15+14)
MultiPolygonTwoCircles/n=10000-4                              5.71ms ±11%    5.70ms ±16%     ~     (p=0.867 n=14+13)
MultiPolygonMultipleTouchingPoints/n=1-4                      4.75µs ±12%    4.58µs ± 6%     ~     (p=0.176 n=14+12)
MultiPolygonMultipleTouchingPoints/n=10-4                     36.6µs ±12%    36.0µs ±10%     ~     (p=0.243 n=13+13)
MultiPolygonMultipleTouchingPoints/n=100-4                     429µs ± 8%     428µs ± 5%     ~     (p=0.981 n=14+13)
MultiPolygonMultipleTouchingPoints/n=1000-4                   4.95ms ± 5%    5.11ms ±12%     ~     (p=0.094 n=13+14)
WKTParsing/point-4                                            1.69µs ± 9%    1.64µs ± 6%     ~     (p=0.148 n=14+12)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           49.4µs ± 8%    50.3µs ±16%     ~     (p=0.621 n=14+15)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            49.5µs ±14%    49.0µs ± 8%     ~     (p=0.793 n=13+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           750µs ± 7%     741µs ± 4%     ~     (p=0.259 n=14+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            753µs ±12%     755µs ± 8%     ~     (p=0.555 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.58µs ±16%    5.44µs ± 8%     ~     (p=0.451 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      5.39µs ±12%    5.53µs ± 9%     ~     (p=0.066 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    60.9µs ± 9%    59.4µs ± 8%     ~     (p=0.202 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     59.9µs ±10%    58.7µs ±12%     ~     (p=0.412 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=100-4                52.9µs ± 9%    52.6µs ±10%     ~     (p=0.943 n=13+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                539µs ± 5%     558µs ± 9%     ~     (p=0.075 n=12+15)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-4                          48.2µs ± 4%    49.9µs ±18%     ~     (p=0.363 n=13+15)
IntersectionWithoutValidation/n=100-4                         93.2µs ± 5%    97.1µs ±26%     ~     (p=0.477 n=14+15)
IntersectionWithoutValidation/n=1000-4                         396µs ±11%     396µs ±16%     ~     (p=0.847 n=14+15)
IntersectionWithoutValidation/n=10000-4                       3.46ms ± 6%    3.70ms ±18%     ~     (p=0.183 n=12+15)
NoOp/n=10-4                                                   3.81µs ± 7%    3.93µs ±18%     ~     (p=0.201 n=15+14)
NoOp/n=100-4                                                  12.3µs ± 9%    12.9µs ±22%     ~     (p=0.125 n=14+14)
NoOp/n=1000-4                                                 87.5µs ± 7%    91.9µs ±16%     ~     (p=0.085 n=14+14)
NoOp/n=10000-4                                                 971µs ±20%     909µs ±26%     ~     (p=0.246 n=14+14)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
SetOperation/n=4/Go_Intersection-4                            41.1µs ±17%    42.1µs ±18%     ~     (p=0.345 n=15+15)
SetOperation/n=4/Go_Difference-4                              41.7µs ± 5%    40.4µs ± 4%   -3.14%  (p=0.006 n=14+13)
SetOperation/n=4/Go_SymmetricDifference-4                     53.8µs ±10%    54.8µs ±16%     ~     (p=0.351 n=15+15)
SetOperation/n=4/Go_Union-4                                   43.6µs ±16%    43.3µs ±11%     ~     (p=0.715 n=14+15)
SetOperation/n=4/GEOS_Intersection-4                          48.5µs ±10%    47.5µs ± 4%     ~     (p=0.125 n=13+13)
SetOperation/n=4/GEOS_Difference-4                            50.5µs ±10%    50.3µs ± 7%     ~     (p=0.880 n=15+14)
SetOperation/n=4/GEOS_SymmetricDifference-4                   69.6µs ± 5%    72.3µs ±15%     ~     (p=0.061 n=13+14)
SetOperation/n=4/GEOS_Union-4                                 53.8µs ±26%    50.1µs ± 4%     ~     (p=0.118 n=15+13)
SetOperation/n=8/Go_Intersection-4                            50.5µs ±14%    49.9µs ±15%     ~     (p=0.252 n=14+15)
SetOperation/n=8/Go_Difference-4                              49.9µs ±10%    49.4µs ± 9%     ~     (p=0.201 n=13+15)
SetOperation/n=8/Go_SymmetricDifference-4                     67.9µs ±10%    67.4µs ±13%     ~     (p=0.591 n=14+15)
SetOperation/n=8/Go_Union-4                                   52.1µs ± 7%    53.0µs ±15%     ~     (p=0.482 n=14+14)
SetOperation/n=8/GEOS_Intersection-4                          56.3µs ± 9%    55.8µs ± 4%     ~     (p=0.840 n=13+13)
SetOperation/n=8/GEOS_Difference-4                            57.8µs ± 8%    56.6µs ± 6%     ~     (p=0.400 n=15+14)
SetOperation/n=8/GEOS_SymmetricDifference-4                   82.2µs ± 6%    81.8µs ± 6%     ~     (p=0.964 n=15+13)
SetOperation/n=8/GEOS_Union-4                                 58.6µs ±10%    58.4µs ± 3%     ~     (p=0.572 n=14+11)
SetOperation/n=16/Go_Intersection-4                           67.2µs ± 7%    71.2µs ±18%     ~     (p=0.112 n=14+15)
SetOperation/n=16/Go_Difference-4                             71.8µs ± 7%    75.0µs ±16%     ~     (p=0.234 n=14+15)
SetOperation/n=16/Go_SymmetricDifference-4                     101µs ± 5%     105µs ±21%     ~     (p=0.847 n=14+15)
SetOperation/n=16/Go_Union-4                                  76.2µs ±12%    79.7µs ±15%     ~     (p=0.137 n=15+15)
SetOperation/n=16/GEOS_Intersection-4                         61.6µs ± 7%    64.7µs ±21%     ~     (p=0.275 n=14+15)
SetOperation/n=16/GEOS_Difference-4                           68.3µs ± 8%    70.8µs ±20%     ~     (p=0.285 n=15+15)
SetOperation/n=16/GEOS_SymmetricDifference-4                   107µs ±15%     107µs ± 5%     ~     (p=0.847 n=15+14)
SetOperation/n=16/GEOS_Union-4                                71.6µs ±16%    71.0µs ±16%     ~     (p=0.505 n=14+15)
SetOperation/n=32/Go_Intersection-4                            118µs ± 8%     115µs ± 7%     ~     (p=0.202 n=14+13)
SetOperation/n=32/Go_Difference-4                              127µs ±15%     121µs ± 6%     ~     (p=0.158 n=15+14)
SetOperation/n=32/Go_SymmetricDifference-4                     172µs ±12%     169µs ±12%     ~     (p=0.233 n=15+15)
SetOperation/n=32/Go_Union-4                                   127µs ±10%     130µs ± 7%     ~     (p=0.201 n=14+15)
SetOperation/n=32/GEOS_Intersection-4                         76.2µs ±15%    73.6µs ± 2%     ~     (p=0.164 n=15+11)
SetOperation/n=32/GEOS_Difference-4                           83.0µs ±11%    82.2µs ± 5%     ~     (p=0.983 n=14+15)
SetOperation/n=32/GEOS_SymmetricDifference-4                   145µs ±14%     142µs ± 9%     ~     (p=0.847 n=14+15)
SetOperation/n=32/GEOS_Union-4                                88.3µs ± 7%    89.2µs ± 6%     ~     (p=0.467 n=13+15)
SetOperation/n=64/Go_Intersection-4                            205µs ±17%     195µs ± 9%   -5.09%  (p=0.025 n=15+13)
SetOperation/n=64/Go_Difference-4                              223µs ±16%     213µs ± 9%     ~     (p=0.389 n=15+15)
SetOperation/n=64/Go_SymmetricDifference-4                     304µs ±13%     293µs ± 8%     ~     (p=0.074 n=15+15)
SetOperation/n=64/Go_Union-4                                   231µs ±18%     224µs ± 5%     ~     (p=0.715 n=15+14)
SetOperation/n=64/GEOS_Intersection-4                         98.5µs ±13%    99.3µs ±14%     ~     (p=0.813 n=14+15)
SetOperation/n=64/GEOS_Difference-4                            121µs ± 9%     120µs ± 8%     ~     (p=0.804 n=14+14)
SetOperation/n=64/GEOS_SymmetricDifference-4                   218µs ±10%     216µs ± 5%     ~     (p=0.756 n=14+13)
SetOperation/n=64/GEOS_Union-4                                 133µs ± 8%     133µs ±15%     ~     (p=0.847 n=15+14)
SetOperation/n=128/Go_Intersection-4                           369µs ± 7%     368µs ±14%     ~     (p=0.624 n=15+15)
SetOperation/n=128/Go_Difference-4                             379µs ± 5%     383µs ±10%     ~     (p=0.734 n=14+14)
SetOperation/n=128/Go_SymmetricDifference-4                    534µs ± 9%     554µs ±12%     ~     (p=0.067 n=15+15)
SetOperation/n=128/Go_Union-4                                  405µs ±10%     403µs ± 6%     ~     (p=0.856 n=15+13)
SetOperation/n=128/GEOS_Intersection-4                         149µs ± 6%     147µs ± 6%     ~     (p=0.375 n=14+13)
SetOperation/n=128/GEOS_Difference-4                           178µs ±12%     180µs ± 7%     ~     (p=0.401 n=14+14)
SetOperation/n=128/GEOS_SymmetricDifference-4                  342µs ± 7%     339µs ± 4%     ~     (p=0.720 n=14+13)
SetOperation/n=128/GEOS_Union-4                                197µs ±11%     194µs ± 9%     ~     (p=0.400 n=15+14)
SetOperation/n=256/Go_Intersection-4                           653µs ± 7%     659µs ± 6%     ~     (p=0.574 n=13+12)
SetOperation/n=256/Go_Difference-4                             721µs ± 7%     712µs ±10%     ~     (p=0.264 n=13+13)
SetOperation/n=256/Go_SymmetricDifference-4                   1.07ms ±14%    1.05ms ± 5%     ~     (p=0.375 n=14+13)
SetOperation/n=256/Go_Union-4                                  780µs ±10%     742µs ± 5%   -4.77%  (p=0.012 n=14+14)
SetOperation/n=256/GEOS_Intersection-4                         219µs ± 3%     219µs ± 9%     ~     (p=0.137 n=14+14)
SetOperation/n=256/GEOS_Difference-4                           300µs ±18%     292µs ± 7%     ~     (p=0.389 n=15+15)
SetOperation/n=256/GEOS_SymmetricDifference-4                  638µs ± 8%     645µs ±12%     ~     (p=0.856 n=13+15)
SetOperation/n=256/GEOS_Union-4                                329µs ± 5%     323µs ± 4%     ~     (p=0.069 n=14+14)
SetOperation/n=512/Go_Intersection-4                          1.34ms ±13%    1.32ms ± 4%     ~     (p=0.511 n=14+14)
SetOperation/n=512/Go_Difference-4                            1.43ms ±10%    1.42ms ±16%     ~     (p=0.325 n=15+15)
SetOperation/n=512/Go_SymmetricDifference-4                   2.01ms ±15%    2.00ms ±14%     ~     (p=0.555 n=13+15)
SetOperation/n=512/Go_Union-4                                 1.46ms ± 5%    1.54ms ±18%     ~     (p=0.063 n=14+15)
SetOperation/n=512/GEOS_Intersection-4                         402µs ± 8%     397µs ± 9%     ~     (p=0.400 n=15+14)
SetOperation/n=512/GEOS_Difference-4                           537µs ±15%     524µs ±10%     ~     (p=0.265 n=14+14)
SetOperation/n=512/GEOS_SymmetricDifference-4                 1.17ms ± 7%    1.17ms ± 5%     ~     (p=0.728 n=13+12)
SetOperation/n=512/GEOS_Union-4                                588µs ± 6%     586µs ± 8%     ~     (p=0.769 n=13+12)
SetOperation/n=1024/Go_Intersection-4                         2.56ms ±15%    2.69ms ±16%     ~     (p=0.116 n=15+15)
SetOperation/n=1024/Go_Difference-4                           2.78ms ± 9%    2.89ms ±14%     ~     (p=0.146 n=15+14)
SetOperation/n=1024/Go_SymmetricDifference-4                  3.99ms ± 4%    4.05ms ±13%     ~     (p=0.936 n=12+13)
SetOperation/n=1024/Go_Union-4                                2.95ms ±12%    2.92ms ± 7%     ~     (p=0.905 n=14+13)
SetOperation/n=1024/GEOS_Intersection-4                        733µs ± 5%     747µs ±15%     ~     (p=0.769 n=14+14)
SetOperation/n=1024/GEOS_Difference-4                         1.05ms ± 5%    1.06ms ± 8%     ~     (p=0.804 n=14+14)
SetOperation/n=1024/GEOS_SymmetricDifference-4                2.34ms ± 3%    2.33ms ± 3%     ~     (p=0.608 n=11+12)
SetOperation/n=1024/GEOS_Union-4                              1.24ms ±11%    1.22ms ±10%     ~     (p=0.477 n=15+14)
SetOperation/n=2048/Go_Intersection-4                         5.80ms ±24%    5.56ms ±13%     ~     (p=0.252 n=15+14)
SetOperation/n=2048/Go_Difference-4                           5.93ms ±15%    5.86ms ± 8%     ~     (p=0.701 n=14+14)
SetOperation/n=2048/Go_SymmetricDifference-4                  8.14ms ± 6%    8.20ms ± 7%     ~     (p=0.579 n=13+13)
SetOperation/n=2048/Go_Union-4                                6.44ms ±20%    6.10ms ± 9%     ~     (p=0.254 n=15+13)
SetOperation/n=2048/GEOS_Intersection-4                       1.48ms ± 4%    1.49ms ±11%     ~     (p=0.940 n=12+14)
SetOperation/n=2048/GEOS_Difference-4                         2.02ms ±22%    1.94ms ±11%     ~     (p=0.983 n=14+15)
SetOperation/n=2048/GEOS_SymmetricDifference-4                4.50ms ±10%    4.56ms ± 7%     ~     (p=0.347 n=12+15)
SetOperation/n=2048/GEOS_Union-4                              2.32ms ±23%    2.27ms ±12%     ~     (p=0.813 n=14+15)
SetOperation/n=4096/Go_Intersection-4                         11.7ms ±14%    11.5ms ±20%     ~     (p=0.512 n=15+15)
SetOperation/n=4096/Go_Difference-4                           12.5ms ±17%    12.5ms ±16%     ~     (p=0.870 n=15+15)
SetOperation/n=4096/Go_SymmetricDifference-4                  18.0ms ±17%    17.9ms ±18%     ~     (p=0.838 n=15+15)
SetOperation/n=4096/Go_Union-4                                12.9ms ±11%    13.5ms ±17%     ~     (p=0.130 n=13+15)
SetOperation/n=4096/GEOS_Intersection-4                       2.81ms ±21%    2.75ms ±19%     ~     (p=0.270 n=15+14)
SetOperation/n=4096/GEOS_Difference-4                         4.19ms ±14%    4.16ms ± 7%     ~     (p=0.801 n=13+13)
SetOperation/n=4096/GEOS_SymmetricDifference-4                9.91ms ±13%   10.20ms ±19%     ~     (p=0.685 n=13+14)
SetOperation/n=4096/GEOS_Union-4                              4.69ms ± 4%    4.98ms ±18%     ~     (p=0.302 n=13+14)
SetOperation/n=8192/Go_Intersection-4                         23.8ms ±22%    23.9ms ±17%     ~     (p=0.769 n=14+14)
SetOperation/n=8192/Go_Difference-4                           25.5ms ±11%    26.3ms ±29%     ~     (p=0.935 n=15+15)
…